Extract ConnectionManager into another file
